DTSPB-2231: Move resetValues function
