+++ app/wrappers/Will.js
@@ -5,33 +5,6 @@ class Will {
         this.will = will || {};
     }
 
-    resetValues(ctx) {
-        if (ctx.willDamageReasonKnown) {
-            ctx.willDamageReasonKnown = 'optionNo';
-        }
-
-        if (ctx.willDamageReasonDescription) {
-            ctx.willDamageReasonDescription = '';
-        }
-
-        if (ctx.willDamageCulpritKnown) {
-            ctx.willDamageCulpritKnown = 'optionNo';
-        }
-
-        if (ctx.willDamageCulpritName) {
-            ctx.willDamageCulpritName = {};
-        }
-
-        if (ctx.willDamageDateKnown) {
-            ctx.willDamageDateKnown = 'optionNo';
-        }
-
-        if (ctx.willDamageDate) {
-            ctx.willDamageDate = '';
-        }
-        return ctx;
-    }
-
     hasCodicils() {
         return this.will.codicils === 'optionYes';
     }
